Added skybox and game logic
Added a plane with a skybox tex that follows the player around with an offset, added score whenever player passes the walls and a difficulty curve that corrosponds to the score.
Also abstracted the walls from cubes to add the Rand() translation to it.
Abstracted Player from cube aswell to remove player code from GameObject (parent) - among other small tweaks...